repair restart handling when running crashes through coredumpd
restarting was originally disabled here because kcrash didn't have the
required metadata. now it does and drkonqi gets invoked accordingly with
--restarted or not.
i.e. we can simply rely on DrKonqi::isRestarted

there may be a future concern here when we implement handling for past
crashes that happened at a previous logout. notably we might opt to
never set --restarted, but not sure it makes a big difference at all