add trivial position tracking to the details text edit
this allow us to be smart and put the cursor where it makes the most
sense, rather than at the beginning or the end.

this way the cursor is positioned at the first "block" of text, that is
a header line related to options the user chose earlier in the report
dialog about what they know about this crash.

it's hardly any more code while also cleaning up the extra newline
handling (which is now happening conditionally on whether there is a
previous block already)

BUG: 438736
FIXED-IN: 5.23