repair query construction
this regressed in d657b0e6023cac063d91a5efb36135f1b6ce7359 when the
qurlquery was changed to qmap. qmap is illsuited as we need keys to
appear multiple times. specifically the duplicate search relies on this
to query both the main product as well as base library products

the end result here was that only one of the many products would get
searched (and usually not the main product either :|)