put dupe detection backtrace parsing in a qfuture
this previously ran on the GUI thread but is potentially a costly
operation as we parse the entire data set all at once (as opposed to the
live tracing as that is done as new lines arrive). this then meant that
the GUI was intermittently locking up when dealing with particularly
egregious traces (or generally very long comments really).

to remove this problem we'll simply run the parsing in a thread on the
threadpool as thankfully the parser is one of the few classes that
doesn't rely on the mountain of singletons we have.