report the drkonqi backend
since coredump is post-mortem debugging it has different capabilities
than the just-in-time debugging of kcrash. e.g. there's a pending merge
request to add qml tracing. but that can only work with the kcrash
backend. with that in mind it's helpful to know the backend to be able
to tell if data is missing because of the backend or because of a bug.

there are also some malfunctions that can appear with JIT debugging that
may lead to bogus reports. e.g. while we handle a crash another thread
might also crash thereby messing up the trace or pointing at a cause
that isn't actually the real reason the app crashed in the first place.