update kdeplatform
half the stuff in the list never was a platform kind of dependency that
would be used by most or all apps.
the other half is from kdelibs4 times.

replace the lot with kcoreaddons. it's realistically the only worthwhile
platform dependency we can expect. while an app could get by without it,
it'd not have kcrash support cause currently kcrash only sends us the KDE
bug address when it finds kaboutdata and kaboutdata are in kcoreaddons.

technically I suppose the more appropriate replacement for kdelibs would be
to list all the frameworks, but ultimately that would bring down performance
and be a lot of work (or require refactoring to how the mappings work).
all in all not worth it.

BUG: 418538