group dolphin in filemanagement and extend that group
dolphin is filemanagement just as much as konqueror. also both are built
on frameworks-kio and indeed KIO is the most likely place for crashes
outside the individual apps

the design of the mappings file is actually a bit fishy. why would we
manually map products to groups but then the groups still contain all
the products that map to it anyway. fairly daft.