remove the 'manually enter bug id' feature
it has an incredibly niche use case to begin with. but what's more it's
also superfluous and sports weird UX.

things wrong with it:

- bug doesn't get added to the model (would cost at least 20 extra lines
of code)
- it takes a whooping 6 extra clicks to get the bug added at all (+API
request delay to fetch the bug and comments)
- the UX is the same as for when you click an arbitrary bug, but of
course this isn't an arbitrary bug. at the point that the reporter can
give drkonqi a bug number the report will already know if this is a
complete dupe or possible dupe, also they will have looked at the bug
and comments already
- being so incredibly useless the entry takes up space that could be
used for a real actually useful dupe candidate
- it needlessly complicates the code because of how it is implemented.
with offsetting falling you on the head in every other function because
the first entry isn't a real entry

the legit use case behind this is that a triager could try to get a
better trace and attach it to the bug. for that it indeed makes a lot of
sense... if it wasn't for the fact that the report can more easily copy
paste the trace manually! having to go through the reporter dialog makes
no sense for this use case: you have to click through numerous wizard
pages. some of them don't even let you move on unless you say you know
what's going on. then you have to go through the 6 extra click nonesense
to mark the bug as dupe. for **zero** extra value added over using the
the main dialog and clicking on the 'developer info' tab and copy
pasting the trace to the bug report. In point of fact, whenever I've
seen a triage do something like this, they've taken the copy paste
route, not the overly complex and annoying reporter route.

and to top it all up the entire thing only even makes sense iff dupe
detection does not work to begin with. i.e. if the dupe finder finds the
dupe, there is no point in manually adding it at all!

BUG: 317558