split apart the dupe page classes into files
every time I opened the dupes file to do anything I got annoyed with how
difficult it is to find anything. the info dialog isn't exactly trivial
and there's lots of name overlap between the various classes so
grepping/searching was a nightmare too.

by having them each be its own file they are easier to digest and work
on