repair backtrace->loginpage shortcut
this repairs a defect where the wizard would get stuck in a loop because
the backtrace->loginpage shorcut would instead go to the versionpage (at
the beginning of the dialog) because the page is always appropriate

I'm not quite sure when this broke (sometime after 5.22 though), or why
it hasn't appeared earlier but this was quite wrong.

our pages have a dedicated 'appropriate' system distinct from the
kassistantdialog system because the latter would manage buttons for us
but we don't want that. as such `KAD::isAppropriate(...)` is never a
function that ought to have been called, yet we did in the version/login
shortcut which then lead to version always being appropriate despite
never actually being.

ideally that shortcut should go away in favor of relying on page
appropriateness but for now let's do the least impactful change and...

- delete isAppropriate so it's less likely to get called by accident
(can still be called through base class but AFAIK there's nothing we can
do about that in c++)
- add a new helper function isItemAppropriate which does much the same
but relying on our appropriateness system
- drop the shortcut to version because it is actually wrong. versionpage
has been moved to the very beginning