separate the wallet entry names when in test mode
it's supremely annoying that test (bugstest.kde.org) and non-test
(bugs.kde.org) mode use the same wallet entries because I have different
credentials on both for testing complex password encoding. instead use a
separate entry name when test mode is enabled

to facilitate this change the static kWalletEntryName was made a member
to set it based on runtime context. also gets rid of the repeated
qlatin1string construction, so I guess that's neat too.

ther's also a neat helper function to hide the details of how to find
out if the testing bugzilla is used