port conclusion page to isApproprite system
this gets rid of hacky shortcuts to skip pages. instead we now have a
page sequence that gets moved through without jumping over stuff. if a
conclusion is not applicable (because the data is good enough) the
conclusion pages are skipped.
currently the same page is modeled by multiple items (one per conclusion
point). it's not super ideal, compared to having separate conclusion
pages with limited logic each, but it's better than what we had before
anyway.

this should make it much harder to get into a page state mess moving
forward