repair the distro selection
this actually broke before 5.21 even but nobody would ever notice
because in absence of a value we had internal fallbacks in place to
ensure a platform is selected. what apparently happend was that until
the recent paging refactor the distro combobox was always hidden by the
time abouttohide ran and so the entire code path was effectively skipped
meaning the distro selection as a whole was kaput (I've reproduced as
much). with the recent refactor the code path became working again and
fell into a trap that had been there for a while:

we actually place the field values in the combobox as text not data, so
what would happen here is that we'd get an invalid qvariant for data,
resulting in an empty string, resulting in an invalid bug POST to the
API, resulting in the API rejecting the submission.

simply use the text instead.
side effect of this fix is that distro selection works again :shrug:

an argument might be made that this combobox as a whole is fairly
useless. if we add a new product to bugzilla we should also add it to
drkonqi and if we didn't then the combobox won't help anyway as the
majority of reporters surely won't pay attention or can't be bothered to
find the right platform in this huge list of stuff

(make the line const correct while at it)

BUG: 442333