replace restart checkbox with button
for one, this aligns the UX a bit more between crash dialog and report
dialog (the former sports a button, not a checkbox, indeed I'm not sure
why one would want this to be a checkbox). it also removes a bug where
restart-on-finish would not run quick enough to finish before the
application exits, since restarting is an async option because we don't
know how long it really takes - instead of rearchitecturing how app exit
works so it can wait for the restart, it seems more reasonable to simply
have the restart action be a button and the user decide when they want
to stop waiting for the restart to happen. since the button kguititem is
now used in two dialogs it has moved to the globals stuff where there is
a new namespace that mimics KStandardGuiItem but for our guiitems.

visualization logic has also been simplified. with the restart action
now being a button with fairly generic text we can show it regardless of
context so long as the app hadn't been restarted.

BUG: 443304