fix the notifier dump truck
this broke a tad in the refactor towards standalone launcher. for one
knotification requires actual eventlooping so the plugin now executes
the qcoreapplication to get that going.

additionally it no longer uses qprocess to start konsole but KIO's
terminallauncherjob. this has the advantage of not hardcoding konsole
and also survives systemd cleaning up. previously the detatched qprocess
would get killed when we had exited because it was part of our cgroup
that gets cleaned up by systemd when the launcher unit stops.
this technically could have used exec() on the job but nested looping is
always garbage but even more so considering knotification has automatica
cleanup facilities that we definitely do not want to run while we are in
a slot. in any event its better to forcefully timeout the start in case
it gets hung on something anyway.

deprecations have also been bumped ahead to get access to the modern
activated signal without having to cast away the ambiguity