bring back idle detection and rejigger timer code
timer logic is now a tad more complicated because there are multiple
scenarios that may occur. to isolate the logic from SNI management at
large there is a new class `ActivationCloseTimer` that contains all the
necessary bits to do the timings.

to also facilitate testing for the logic the new timer class is
dependency inverted from kidletime and dbus allowing us to insert duds
and not have to set up exhaustive test environments. it sucks a bit
though because it also means the actual watchers aren't tested :|