new tool drkonqi-coredump-gui
it is a GUI for coredumpctl - duh.

There are two main motivations behind this

- provide more user friendly access to coredumps (and thus allow more
crashes to be reported to the relevant authors)
- it is the foundation for plans to aggregate crashes. right now
multiple crashes result in multiple drkonqis but realistically we could
just feed them all into a singular UI. coredump-gui kind of gets that
started as the UI bits actually aren't coredump related at all so we can
iterate from here

this also contains some rejiggering of the existing coredump code to
facilitate more code sharing through a new internal static library.

architecture-wise the gui is super simple. there's a central Patient
concept which "abstracts" a crashed application. Patients are managed by
a PatientModel (which is really just a generic QObject model) and get
fed into that model by the existing coredumpwatcher tech. the qml bits
then simply view the model/patients. without coredumpd in the picture
the Patients would be synthesized through other means as yet to be
invented.