finalize and follow the rule of five
since we need a dtor here we need to tag it and the Product isn't
expected to be derived so we can mark it final for the time being. this
isn't library code so it can be changed in the future if need be.
HTTPConnection meanwhile could be derived so let's leave it override.

also since we have a dtor we want to follow the rule of five and
implement copy and move ctors -> disable them explicitly