allow using debuginfod for symbol resolution with gdb12
gdb defaults to it being off.

this adds a new cmake option WITH_GDB12 to be set by distros. when this
is set (and the gdb debugger is used by drkonqi at runtime) then the
install debug symbols button gets slightly refined behavior where
instead of going to a separate tool to install symbols it will instead
run gdb with a suitable cmdline to enable debuginfod.

we still prefer the separate installer when also available.

in particular the workflow here looks like this:

- crash & user opens drkonqi
- trace happens without debuginfod
- if the trace is poor the install debug info button gets enabled
- user clicks button
- re-tracing with debuginfod enabled
- trace is now amazing and can be reported

BUG: 454063