sentry-data: new util to fetch data for preamble.py
this replaces the shoddy qdbus invocations and more generally will serve
as a place to collect data that are difficult to get a hold of in python
(e.g. dbus, or opengl stuff)

since the tracing runs in a different process form the gui we don't need
to worry about blocking dbus calls here. still, lower the dbus timeouts
a tad to not hold up the tracing for multiple minutes