libbugzilla: port from kio to qnam
kio just proofs way to unreliable for library use cases since it was
designed around browser use cases (e.g. can been seen by the metadata
workarounds we had going on) I also have the sneaking suspicion that it
was the root of various submission-getting-stuck problems people
occasionally have but without much information to go on.

to avoid major changes to the API the qnam lives inside the job, even
though conceptually it's more appropriate in the connection. trouble is
that the connection get/put/post are const while the qnams are not so
there is a bit of a conflict there