enable coredump processor & launcher & cleanup by default
this rejiggers the new systemctlenable code a tad so it can be used for
both system and user services

coredump processors jumps into action when a coredump is collected and
the crash hadn't yet been dealt with by drkonqi. it will then process
the dump and potentially invoke the user scoped launcher to start
drkonqi.
the cleanup tech makes sure we don't litter kcrash metadata