coredump-processor: new pickup behavior mode
pickup mode is meant to be run as user and filters out user specific
crashes to then run the launcher on. it only runs on login and is meant
to pick up crashes that happened during logout/shutdown.

the previous shutdown report logging is considered replaced by coredumpd
and has been removed