preamble: don't trip over bad mappings
log an error in sentry and move on with the rest of the preamble.
ideally this should give us actual reports for the defects that had the
mapping fail. as of right now it's very unclear why mappings happen to
fail unexpectedly (e.g. a plasmashell crash that only has plasmashell
itself as objfiles but info proc mappings shows more files loaded -- so
far I've been unsuccessful in reproducing this)