sentry: add ability to send user messages along
this actually was an existing feature but limited to run as part of the
bug reporting workflow. since that workflow is now fairly hidden we need
a new way to get data from the user.

the sentrypage now contains a feedback field.

this required some reshuffling on the backend side of things since we
have two scenarios to consider: comment finished before trace, and trace
finished before comment. this is insofar a complication as currently
sentry feedback payloads must be associated with an event, so the only
legit ways of transmitting the data is event&feedback or event followed
by feedback, but never feedback followed by event. headache material.

the primary change regarding this is that the event_id from python is
now overridden by an envelope event_id. the latter now provides a stable
id for both event and user_report instead of the previous nonesense of
extracting the event_id from the event to serve as envelope id.

additionally we can now write the same envelope twice. this allows us to
get away with the original UX - the event sends ASAP. a possible
user_report will then be submitted later by simply re-delivering the
entire envelope including user_report (the excess event is discarded
server side).