windows: drop support
I've checked with the kde-windows room and we've conclude that this
probably isn't used. let's remove it and see who comes out of the
woodwork.

this code hasn't really seen any changes making me question it's
quality. at the same time it is very questionable anyway: windows has
it's own crash handling infrastructure, drkonqi is only getting in the
way there.

this is a first step towards reducing the infrastructure code for
multiple debuggers