polkit: correctly call fcntl
bit of a doozy this. fcntl requires a third argument as start position
for finding a free file descriptor. the previous incorrect call happened
to work in debug builds (presumably because the invalid memory would be
0x0) but failed in release builds (presumably because the invalid memory
would be some random large value).

explicitly set the target fd argument to at least 0 since we don't
really care what number the fd is so long as it is valid.

BUG: 487727