implement ANR support
application not responding (ANR) is based on kwin's ability to notice if
a GUI bugs out and stops responding. kwin will then offer the user the
ability to terminate the process. by expanding this feature a bit using
metadata we can thus get fairly reasonable ANR detection.

this is limited to the coredumpd backend because we need some metadata
we don't necessarily get from the kcrash backend.

crashedapplication has grown a new wasNotResponding property which
indicates the ANR detection. this is false by default but may be
switched to true if a kwin ANR marker is found for the crash

if the crash was ANR we'll massage the report data to conform to other
platform's ANR reports. this notably includes that sentry reports the
crash as "ApplicationNotResponding"