drop ptracer support
this was used to pass debugging on to external applications but that
doesn't work anymore with coredumpd based crashes and it was a niche
feature one had to opt into anyway. it also never was ported into the
qml ui so it has been effectively removed for a while now and nobody
noticed or complained.