backtracegenerator: cache raw debugger output and give the user access to it
this now means we capture both stdout and stderr. shouldn't be a huge
problem because we ultimately run the output through the gdb parser
anyway and that filters out most of the noise. plus, we mostly care
about the traces in sentry anyway and those are generated by the
preamble and clean.

there are two ways to get the data
- the inlinemessage now sports a details link that opens a txt file
- via the developer view, which no longer shows a useless error

there area also now verbosity lines to mark the beginning and the end of
the debugging session

BUG: 477586
BUG: 474116
BUG: 478110
CCBUG: 491820