backtracegenerator: implement a serialization lock
to prevent systems from drowning in gdb instances, we now establish a
serialization lock in the form of a file lock. only a single debugger
may be active at a time

this only partially fixes the problem since a single debugger may still
OOM the system. we still want a better solution, such as resource
constraining via systemd, in the long run.

CCBUG: 489315