preamble: don't stumble over missing build ids in eu-unstrip
when the library has no build-id proper we need to raise ourselves out
of the mess rather than trying to split the '-'

this may actually need refinement in the future because (e.g.)
proprietary libraries may have no build-id. at the same time they also
cannot be traced though, so perhaps we need to discard the trace only
when they are actually used in a frame?